Incremento de botões para maior interação com usuário.
